+++ osutils/osutils.go
@@ -1,7 +1,6 @@
-package utils
+package osutils
 
 import (
-	"fmt"
 	"io/fs"
 	"os"
 	"runtime"
@@ -26,20 +25,6 @@ var (
 	}
 	defaultOS = Linux
 )
-
-func ByteCountSI(b int64) string {
-	const unit = 1000
-	if b < unit {
-		return fmt.Sprintf("%d B", b)
-	}
-	div, exp := int64(unit), 0
-	for n := b / unit; n >= unit; n /= unit {
-		div *= unit
-		exp++
-	}
-	return fmt.Sprintf("%.1f %cB",
-		float64(b)/float64(div), "kMGTPE"[exp])
-}
 
 func CurrentOS() *MachineOSInfo {
 	if info := reg[runtime.GOOS]; info != nil {
